Pull process_open_files out of processes.cpp and reduce logging
