+++ osquery/tables/specs/x/processes.table
@@ -1,19 +1,19 @@
 table_name("processes")
 schema([
-    Column("name", TEXT),
+    Column("name", TEXT, "The process path or shorthand argv[0]"),
     Column("path", TEXT),
-    Column("cmdline", TEXT),
+    Column("cmdline", TEXT, "Complete argv"),
     Column("pid", INTEGER),
     Column("uid", BIGINT),
     Column("gid", BIGINT),
     Column("euid", BIGINT),
     Column("egid", BIGINT),
-    Column("on_disk", TEXT),
-    Column("wired_size", TEXT),
-    Column("resident_size", TEXT),
-    Column("phys_footprint", TEXT),
-    Column("user_time", TEXT),
-    Column("system_time", TEXT),
+    Column("on_disk", TEXT, "The process path exist yes=1, no=-1"),
+    Column("wired_size", INTEGER),
+    Column("resident_size", INTEGER),
+    Column("phys_footprint", INTEGER),
+    Column("user_time", INTEGER),
+    Column("system_time", INTEGER),
     Column("start_time", TEXT),
     Column("parent", INTEGER),
 ])
